Implement transparent communication between client and server.
The client runs the Scala.js version of Akka, and the server runs
a classic Play server with its embedded Akka. A route in Play
defines a WebSocket entry point to an actor in Akka, and the client
can connect to that entry point and receive an ActorRef for that
actor. From there, it can send pickleable messages to that ActorRef,
and receive replies. ActorRefs are pickleable so they can be put
in messages.