Slightly better workflow for the chat example.
It is not necessary to run explicitly `chatExampleScalaJS/packageJS`
anymore. It is triggered automatically by the `chatExample/compile`
task. That means it is run automatically, when you launch `run`,
but changing sources of the client app will not trigger the
auto-refresh-on-reload mechanism of Play. You still need to stop
the server (Ctrl+D) and launch it again (run), but this is already
much better than it was before.