More principled way of fetching IR patches.
Instead of writing the class directory to a config file, and
*depending* on the `akkaJsActorIrPatches` project, we simply read
`products in (akkaJsActorIrPatches, Compile)`.

In addition to being substandard in terms of sbt task dependency
graph, the previous solution caused the classDirectory of the IR
patches to end up in `akkaJsActor`'s `fullClasspath` (despite the
`% "provided"`). Although this did not have an impact on published
artifacts (AFAICT), it polluted the set of IR files seen in this
build.