Fix the IR checking errors related to setting the vals of `Actor`.
Since `Actor` is really a trait, patching its IR with fields,
getters and setters from a `class` is nonsensical, and produces
invalid IR.

Instead, we simply abuse scalac's internal encoding of `val`s in
traits. For concrete `val`s in traits, scalac generates an internal
special setter, called from the constructor of the trait to
initialize the fields mixed in the subclasses. We simply hijack the
mechanism and call those methods reflectively.

This is dependent on the specific encoding of traits by scalac,
so it could be broken in a future (binary incompatible) version of
Scala. However, for now there seems to be no plan for such changes,
and this fixes the IR errors.